all byte->int (any kind of memsize)

+++ 16bit-machine/src/com/parts/ILogicPart.java
@@ -3,6 +3,6 @@ package com.parts;
 public interface ILogicPart
 {   
     public void computeStep();
-    public void writeToInStream(byte[] in);
-    public byte readByteFromOutStream();
+    public void writeToInStream(int[] in);
+    public int readByteFromOutStream();
 }
